Refactor dialog in profile to appropriate place
